Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "openai>=2.15.0",
    "passlib[argon2,bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]
